Prevent accidental removal of chef-client daemon pidfile when child processes of the daemon exit.
Children forked from the chef-client daemon (such as those triggered
during runs of only_if commands) inherit the at_exit handler of the
daemon, which removes the daemon's pid file.  remove_pid_file now
only removes the pid file if it's invoked by the same process which
created it in the first place.

https://bugzilla.novell.com/show_bug.cgi?id=773182